Convert image into double before segmenting
git-svn-id: https://orchestra.med.harvard.edu/svn/lccb/matlab/projects/common/main@8305 0bcb941b-74d1-0310-bb26-e6636707f451

+++ imageSegmentation/segmentationPackage/getMovieMasksMSS.m
@@ -122,7 +122,7 @@ for i=1:numel(p.ChannelIndex)
         if isempty(p.ProcessIndex)
             currImage = movieData.channels_(iChan).loadImage(j);
         else
-            currImage = movieData.processes_{p.ProcessIndex}.loadOutImage(iChan,j);
+            currImage = double( movieData.processes_{p.ProcessIndex}.loadOutImage(iChan,j) );
         end
         
         mask=getCellMaskMSS(currImage,'Scales',p.Scales,'FilterOrder',p.FilterOrder);
